Refactor driver architecture for better thread safety (#22)
Addresses PR review comments by:
1. Moving hardware driver to a standalone, reusable class.
2. Using a context manager to guarantee threads are joined.
