+++ ros2_hc_drv/pupil_labs_ros/setup.py
@@ -1,30 +1,39 @@
-from setuptools import setup, find_packages
-from glob import glob
 import os
+from glob import glob
 
-package_name = 'pupil_labs_ros'
+from setuptools import find_packages, setup
+
+
+package_name = "pupil_labs_ros"
 
 setup(
     name=package_name,
-    version='0.0.0',
+    version="0.0.0",
     packages=find_packages(),
     data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        ('share/' + package_name, ['package.xml']),
-        (os.path.join('share', package_name, 'launch'), glob('launch/*.launch.py')),
-        (os.path.join('share', package_name, 'config'), glob('config/*.yaml')),
+        (
+            "share/ament_index/resource_index/packages",
+            ["resource/" + package_name],
+        ),
+        ("share/" + package_name, ["package.xml"]),
+        (
+            os.path.join("share", package_name, "launch"),
+            glob("launch/*.launch.py"),
+        ),
+        (
+            os.path.join("share", package_name, "config"),
+            glob("config/*.yaml"),
+        ),
     ],
-    install_requires=['setuptools'],
+    install_requires=["setuptools"],
     zip_safe=True,
-    maintainer='Mohamed Eyad',
-    maintainer_email='mooeyad@gmail.com',
-    description='ROS 2 driver for Pupil Labs Neon glasses',
-    license='MIT',
+    maintainer="Mohamed Eyad",
+    maintainer_email="mooeyad@gmail.com",
+    description="ROS 2 driver for Pupil Labs Neon glasses",
+    license="MIT",
     entry_points={
-        'console_scripts': [
-            'pupil_node = pupil_labs_ros.pupil_node:main',
-            'pupil_labs_neon = pupil_labs_ros.pupil_labs_neon_ros:main',
+        "console_scripts": [
+            "pupil_labs_neon = pupil_labs_ros.pupil_labs_neon_ros:main",
         ],
     },
 )
